chore[react-devtools]: extract ReactWorkTagMap into a separate module
